FAFBFL owl n2n

+++ src/VFB_connectomics_import/flywire_import.py
@@ -2,6 +2,7 @@
 #    Warning: be careful with versions and rootid mappings for this dataset. Contact Philipp Schlegel at FlyWire for up to date advice (suggest using slack)
 
 from fafbseg import flywire
+import fafbseg
 import navis
 import pandas as pd
 import flybrains
@@ -67,63 +68,39 @@ curation_tsv=generate_curation_tsv(annotations_tsv_path)
 curation_tsv.to_csv('Dorkenwald2023_curation.tsv', sep='\t', index=False)
 
 ##### scraps for testing skels/meshes, code runs on jenkins (url tbd).
+
 ## testing neuron obj sizes
-testset = list(curation_tsv.filename[0:100])
-k = flywire.get_mesh_neuron(testset, dataset='flat_630')
-s = flywire.get_skeletons(testset) #only works for 630 so no need to specify dataset
-
-## APL test TODO try aligning as for manc
-m = flywire.get_mesh_neuron(720575940621280688)
-n = flywire.get_skeletons(720575940621280688)
-
-m = flywire.get_mesh_neuron(720575940606577414, dataset='flat_630')
-n = flywire.get_skeletons(720575940606577414)
+testset = list(curation_tsv.filename[10000:10010])
+k = flywire.get_mesh_neuron(testset, dataset='flat_783', lod=2)
+s = flywire.get_skeletons(testset, dataset='783') #there are some neurons with multiple somas?
 
 ## plotting
-navis.plot3d([k[0:10]],backend='plotly').show(renderer='browser')
-
+navis.plot3d([k],backend='plotly').show(renderer='browser')
+navis.plot3d([s],backend='plotly').show(renderer='browser')
 
 #transforming
 flybrains.download_jrc_transforms()
 flybrains.register_transforms()
-transformed_neuron = navis.xform_brain(k[0:10], source='FLYWIRE', target='JRC2018U')
+transformed_neuron = navis.xform_brain(k, source='FLYWIRE', via='JRCFIB2022M' ,target='JRC2018U')
 #plot transformed against mesh
 navis.plot3d([transformed_neuron, flybrains.JRC2018U],backend='plotly').show(renderer='browser')
 
-## writing
-navis.write_mesh(m,filepath='test.obj')
-navis.write_swc(n, filepath='test.swc')
-
-## read swc from local file (should have fixed somata)
-a=navis.read_swc('/Users/adm/Downloads/sk_lod1_630_healed_ds2/720575940621280688.swc')
-
-# test if all of the filenames exist in the local files (using mat 630 ids so should map to the meshes 1:1)
-for neuron in curation_tsv.filename:
-    filepath = '/Users/adm/Downloads/sk_lod1_630_healed_ds2/'+str(neuron)+'.swc'
-    try:
-        navis.read_swc(filepath)
-    except Exception as e: print(e)
 
 
 
-###
+
+
+
+
 #!.pyenv/bin/python3.9
 
 import os
 import glob
-import numpy as np
 import navis
 import flybrains
-import pandas as pd
-import cloudvolume
-import sys
-from vfb_connect.cross_server_tools import VfbConnect
-import multiprocessing
-from multiprocessing import Pool
-import threading
 from fafbseg import flywire
-
-os.getenv('OUTPUT_FOLDER')
+import pandas as pd
+from vfb_connect.cross_server_tools import VfbConnect
 
 def delete_volume_files(local_folder_path):
     """
@@ -133,7 +110,6 @@ def delete_volume_files(local_folder_path):
     # Ensure the path ends with a separator
     if not local_folder_path.endswith(os.sep):
         local_folder_path += os.sep
-
     # Generate file paths for all files matching the pattern "volume.*" in the specified folder
     file_paths = glob.glob(f'{local_folder_path}volume*')
     file_paths += glob.glob(f'{local_folder_path}thumbnail*')
@@ -145,7 +121,7 @@ def delete_volume_files(local_folder_path):
         except Exception as e:
             print(f'Could not delete {file_path}: {e}')
 
-# Download the JRC transforms
+# Download transforms
 flybrains.download_jrc_transforms()
 
 # Register the downloaded transforms
@@ -154,7 +130,7 @@ flybrains.register_transforms()
 # Setup connection to the VFB Neo4j database
 kbw = VfbConnect(neo_endpoint='http://kb.virtualflybrain.org:80', neo_credentials=('neo4j', os.environ.get('password')))
 
-kbw = VfbConnect(neo_endpoint='http://kb.virtualflybrain.org:80')
+flywire.set_chunkedgraph_secret("47576875e49eeb999396a0e73de0cf31")
 
 # Your Cypher query
 cypher_query = '''
@@ -172,7 +148,7 @@ if output is False or not output:
     exit(1)  # Exit the script in case of an error
 
 # Access the data key of the first item in the output list
-query_data_df=pd.DataFrame(output[0]['data'])['row'].apply(pd.Series)
+query_data_df = pd.DataFrame(output[0]['data'])['row'].apply(pd.Series)
 
 # Get the total number of items to process
 total_items = len(query_data_df[0])
@@ -180,9 +156,19 @@ total_items = len(query_data_df[0])
 # Process the results
 for index, result in query_data_df.iterrows():
     print(f"Processing item {index + 1} of {total_items}")
-    # Access the row key of each result item to get the root_id and folder values
-    root_id, folder_url = result
+    # Access the row key of each result item to get the bodyid and folder values
+    body_id, folder_url = result
     folder_url = folder_url.replace('http://www.virtualflybrain.org/data/', '/IMAGE_WRITE/')
+
+    try:
+
+        body_id = int(body_id)  # Ensure body_id is an integer
+
+    except ValueError as e:
+
+        print(f"Warning: body_id {body_id} is not an integer. Error: {e}")
+
+        continue
 
     # Create the local folder if it doesn't exist
     local_folder_path = os.path.dirname(folder_url)
@@ -194,58 +180,88 @@ for index, result in query_data_df.iterrows():
     # Path for the output obj file
     mesh_filename = os.path.join(local_folder_path, f"volume_man.obj")
 
+    # Path for the output NRRD file
+    nrrd_filename = os.path.join(local_folder_path, "volume.nrrd")
+
     # Check if the output file already exists, if so, skip to the next iteration
-    if os.path.exists(filename) and os.path.exists(mesh_filename):
+    if os.path.exists(filename) and os.path.exists(mesh_filename) and os.path.exists(nrrd_filename):
         # Redo files if requested
         if os.environ.get('redo') == 'true':
-            print(f"File for neuron {root_id} swc and obj already exists. Removing old versions...")
+            print(f"File for neuron {body_id} swc and obj already exists. Removing old versions...")
             delete_volume_files(local_folder_path)
         else:
-            print(f"File for neuron {root_id} swc or obj already exists. Skipping...")
-            continue #skipping nrrd for now
-            if os.path.exists(nrrd_filename):
-                continue
+            print(f"File for neuron {body_id} swc or obj already exists. Skipping...")
+            continue
 
     if os.path.exists(filename) or os.path.exists(mesh_filename):
         # Redo files if requested
         if os.environ.get('redo') == 'true':
-            print(f"File for neuron {root_id} swc or obj already exists. Removing old versions...")
+            print(f"File for neuron {body_id} swc or obj already exists. Removing old versions...")
             delete_volume_files(local_folder_path)
 
     try:
         if not os.path.exists(filename):
             # Download skeleton
-            ### this needs to be the path to the local 'sk_lod1_630_healed_ds2' folder + the root_id - ex 'c/sk_lod1_630_healed_ds2/720575940621280688.swc'
-            neuron = navis.read_swc('sk_lod1_630_healed_ds2' + root_id)
+            neuron = flywire.get_skeletons(body_id, dataset='783')
 
-            # Transform neuron data from FlyWire to JRC2018U
-            transformed_neuron = navis.xform_brain(neuron, source='FLYWIRE', target='JRC2018U')
+            #replace the above line if the swcs are downloaded instead from
+            ### this needs to be the path to the local 'sk_lod1_783_healed' folder + the root_id - ex 'c/sk_lod1_783_healed_ds2/720575940621280688.swc'
+            #neuron = navis.read_swc('sk_lod1_783_healed' + root_id)
+
+            # Transform neuron data from FLYWIRE to JRC2018U (needs via to work currently)
+            transformed_neuron = navis.xform_brain(neuron, source='FLYWIRE', via='JRCFIB2022M' ,target='JRC2018U')
             # Check for warnings
             if not transformed_neuron:
-                print(f"Warning: No skeleton data available for neuron {root_id}. Possibly outside the H5 deformation field.")
+                print(
+                    f"Warning: No skeleton data available for neuron {body_id}. Possibly outside the H5 deformation field.")
                 continue  # Skip to the next iteration
             # Save to SWC file
             navis.write_swc(transformed_neuron, filename)
             print(f"Written to {filename.replace('/IMAGE_WRITE/', 'http://www.virtualflybrain.org/data/')}")
         else:
-            print(f"File for neuron {root_id} swc already exists. Skipped...")
+            print(f"File for neuron {body_id} swc already exists. Skipped...")
     except Exception as e:
-        print(f"Warning: Failed to process neuron {root_id}. Error: {e}")
+        print(f"Warning: Failed to process neuron {body_id}. Error: {e}")
     try:
-        if not os.path.exists(mesh_filename):
-            # Download mesh
-            mesh_neuron = flywire.get_mesh_neuron(root_id, dataset='flat_630', omit_failures=True)
+        if not os.path.exists(mesh_filename) or not os.path.exists(nrrd_filename):
+            mesh_neuron = neu.fetch_mesh_neuron(body_id, missing_mesh='warn', lod=None)
 
-            # Transform neuron mesh data from FlyWire to JRC2018U
-            transformed_mesh_neuron = navis.xform_brain(mesh_neuron, source='FLYWIRE', target='JRC2018U')
+            # Transform neuron mesh data from FLYWIRE to JRC2018U (needs via to work currently)
+            transformed_mesh_neuron = navis.xform_brain(mesh_neuron, source='FLYWIRE', via='JRCFIB2022M' ,target='JRC2018U')
             # Check for warnings
             if not transformed_mesh_neuron:
-                print(f"Warning: No mesh data available for neuron {root_id}. Possibly outside the H5 deformation field.")
+                print(
+                    f"Warning: No mesh data available for neuron {body_id}. Possibly outside the H5 deformation field.")
                 continue  # Skip to the next iteration
-            # Save to obj file
-            navis.write_mesh(transformed_mesh_neuron, mesh_filename)
-            print(f"Written to {mesh_filename.replace('/IMAGE_WRITE/', 'http://www.virtualflybrain.org/data/')}")
+            if not os.path.exists(mesh_filename):
+                # Save to obj file
+                navis.write_mesh(transformed_mesh_neuron, mesh_filename)
+                print(f"Written to {mesh_filename.replace('/IMAGE_WRITE/', 'http://www.virtualflybrain.org/data/')}")
         else:
-            print(f"File for neuron {root_id} obj already exists. Skipped...")
+            print(f"File for neuron {body_id} obj already exists. Skipped...")
     except Exception as e:
-        print(f"Warning: Failed to process neuron {root_id}. Error: {e}")
+        print(f"Warning: Failed to process neuron {body_id}. Error: {e}")
+
+    # Check if the output NRRD file already exists, if so, skip to the next iteration
+
+    if os.path.exists(nrrd_filename):
+        print(f"File for neuron {body_id} nrrd already exists. Skipping...")
+        continue
+
+    # Create NRRD via navis
+    try:
+        # Check for warnings
+        if not transformed_mesh_neuron:
+            print(
+                f"Warning: No mesh data available for neuron {body_id}. Possibly outside the H5 deformation field. Trying Skeleton")
+            transformed_mesh_neuron = transformed_neuron
+            if not transformed_mesh_neuron:
+                print(
+                    f"Warning: No skeleton data available for neuron {body_id}. Possibly outside the H5 deformation field.")
+                continue  # Skip to the next iteration
+        vx = navis.voxelize(transformed_mesh_neuron, pitch=['0.5189161 microns', '0.5189161 microns', '1.0 microns'], bounds=[[0, 627.3695649], [0, 293.1875965], [0, 173]], parallel=True)
+        vx.grid = (vx.grid).astype('uint8') * 255
+        navis.write_nrrd(vx, filepath=nrrd_filename, compression_level=9)
+        print(f"Written to {nrrd_filename.replace('/IMAGE_WRITE/', 'http://www.virtualflybrain.org/data/')}")
+    except Exception as e:
+        print(f"Warning: Failed to create NRRD for neuron {body_id}. Error: {e}")
